+++ src/udp/UDPProtocol.h
@@ -29,10 +29,10 @@ public:
 
 	int getHeaderSize() const { return header_size;};
 
-	int32_t getTotalBytes() const { return total_bytes_; };
-	uint64_t getTotalPackets() const { return total_packets_;};
-	uint64_t getTotalValidatedPackets() const { return total_validated_packets_;};
-	uint64_t getTotalMalformedPackets() const { return total_malformed_packets_;};
+	int64_t getTotalBytes() const { return total_bytes_; };
+	int64_t getTotalPackets() const { return total_packets_;};
+	int64_t getTotalValidatedPackets() const { return total_validated_packets_;};
+	int64_t getTotalMalformedPackets() const { return total_malformed_packets_;};
 
         void setMultiplexer(MultiplexerPtrWeak mux) { mux_ = mux; };
         MultiplexerPtrWeak getMultiplexer() { mux_;};
@@ -93,7 +93,7 @@ private:
 	FlowForwarderPtrWeak flow_forwarder_;
 	struct udphdr *udp_header_;
 
-	int32_t total_bytes_;
+	int64_t total_bytes_;
 };
 
 typedef std::shared_ptr<UDPProtocol> UDPProtocolPtr;
